+++ src/main/java/com/jasik/momsnaggingapi/domain/schedule/controller/ScheduleController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.schedule.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ScheduleController {
+}
