+++ src/main/java/com/jasik/momsnaggingapi/domain/schedule/repository/ScheduleRepository.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.schedule.repository;
+
+import com.jasik.momsnaggingapi.domain.schedule.Schedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+}
