+++ src/main/java/com/jasik/momsnaggingapi/domain/schedule/service/ScheduleService.java
@@ -1,0 +1,13 @@
+package com.jasik.momsnaggingapi.domain.schedule.service;
+
+import com.jasik.momsnaggingapi.domain.schedule.repository.ScheduleRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ScheduleService {
+    private final ScheduleRepository scheduleRepository;
+}
